+++ packages/core/lib/src/data/css.dart
@@ -150,15 +150,10 @@ class CssLength {
   final CssLengthUnit unit;
 
   /// Creates a measurement.
-  ///
-  /// [number] must not be negative.
-  const CssLength(
-    this.number, [
-    this.unit = CssLengthUnit.px,
-  ]) : assert(number >= 0);
+  const CssLength(this.number, [this.unit = CssLengthUnit.px]);
 
-  /// Returns `true` if value is non-zero.
-  bool get isNotEmpty => number > 0;
+  /// Returns `true` if value is larger than zero.
+  bool get isPositive => number > 0.0;
 
   /// Calculates value in logical pixel.
   double? getValue(TextStyleHtml tsh,
@@ -245,11 +240,11 @@ class CssLengthBox {
       );
 
   /// Returns `true` if any of the left, right, inline measurements is set.
-  bool get hasLeftOrRight =>
-      _inlineEnd?.isNotEmpty == true ||
-      _inlineStart?.isNotEmpty == true ||
-      _left?.isNotEmpty == true ||
-      _right?.isNotEmpty == true;
+  bool get hasPositiveLeftOrRight =>
+      _inlineEnd?.isPositive == true ||
+      _inlineStart?.isPositive == true ||
+      _left?.isPositive == true ||
+      _right?.isPositive == true;
 
   /// Calculates the left value taking text direction into account.
   double? getValueLeft(TextStyleHtml tsh) => (_left ??
@@ -260,6 +255,40 @@ class CssLengthBox {
   double? getValueRight(TextStyleHtml tsh) => (_right ??
           (tsh.textDirection == TextDirection.ltr ? _inlineEnd : _inlineStart))
       ?.getValue(tsh);
+
+  @override
+  String toString() {
+    const _null = 'null';
+    final left = (_left ?? _inlineStart)?.toString() ?? _null;
+    final top = this.top?.toString() ?? _null;
+    final right = (_right ?? _inlineEnd)?.toString() ?? _null;
+    final bottom = this.bottom?.toString() ?? _null;
+    if (left == right && right == top && top == bottom) {
+      return 'CssLengthBox.all($left)';
+    }
+
+    final values = [left, top, right, bottom];
+    if (values.where((v) => v == _null).length == 3) {
+      if (left != _null) {
+        if (_left != null) {
+          return 'CssLengthBox(left=$_left)';
+        } else {
+          return 'CssLengthBox(inline-start=$_inlineStart)';
+        }
+      }
+      if (top != _null) return 'CssLengthBox(top=$top)';
+      if (right != _null) {
+        if (_right != null) {
+          return 'CssLengthBox(right=$_right)';
+        } else {
+          return 'CssLengthBox(inline-end=$_inlineEnd)';
+        }
+      }
+      if (bottom != _null) return 'CssLengthBox(bottom=$bottom)';
+    }
+
+    return 'CssLengthBox($left, $top, $right, $bottom)';
+  }
 }
 
 /// Length measurement units.