+++ packages/core/lib/src/internal/ops/style_padding.dart
@@ -24,7 +24,7 @@ class StylePadding {
         onTree: (meta, tree) {
           if (meta.willBuildSubtree == true) return;
           final padding = tryParseCssLengthBox(meta, kCssPadding);
-          if (padding == null || !padding.hasLeftOrRight) return;
+          if (padding == null || !padding.hasPositiveLeftOrRight) return;
 
           return wrapTree(
             tree,
@@ -41,25 +41,30 @@ class StylePadding {
           final padding = tryParseCssLengthBox(meta, kCssPadding);
           if (padding == null) return null;
 
-          return listOrNull(wf
-              .buildColumnPlaceholder(meta, widgets)
-              ?.wrapWith((c, w) => _build(c, meta, w, padding)));
+          return [
+            WidgetPlaceholder(
+              padding,
+              child: wf.buildColumnPlaceholder(meta, widgets),
+            ).wrapWith(
+              (context, child) => _build(meta, context, child, padding),
+            )
+          ];
         },
         onWidgetsIsOptional: true,
         priority: kPriorityBoxModel3k,
       );
 
-  Widget? _build(BuildContext context, BuildMetadata meta, Widget child,
+  Widget? _build(BuildMetadata meta, BuildContext context, Widget child,
       CssLengthBox padding) {
     final tsh = meta.tsb.build(context);
     return wf.buildPadding(
       meta,
       child,
       EdgeInsets.fromLTRB(
-        padding.getValueLeft(tsh) ?? 0,
-        padding.top?.getValue(tsh) ?? 0,
-        padding.getValueRight(tsh) ?? 0,
-        padding.bottom?.getValue(tsh) ?? 0,
+        max(padding.getValueLeft(tsh) ?? 0.0, 0.0),
+        max(padding.top?.getValue(tsh) ?? 0.0, 0.0),
+        max(padding.getValueRight(tsh) ?? 0.0, 0.0),
+        max(padding.bottom?.getValue(tsh) ?? 0.0, 0.0),
       ),
     );
   }